Move "real time" animation code from DefaultAnimationLoop to GUI code
DefaultAnimationLoop was changed in d4c1c155 to run in "real time" when
called with a 'dt' value of zero, but this conflicts with another
behavior: when called with a 'dt' of 0, Simulation::animate() is
supposed to take the 'dt' parameter from the graph.  As a result, the
'dt' attribute from the root node was sometimes ignored, for example in
batch mode.

To fix this issue, this commit restores the normal behavior of
DefaultAnimationLoop and moves the "real time" code to the GUI code.  It
also adds a "Real Time" checkbox in the GUI of runSofa to make this
feature more visible.

This effectively reverts d4c1c155 and 30803af6, and moves the changes
introduced by those commits from DefaultAnimationLoop to RealGUI.{h,cpp}


Former-commit-id: e3a3136b17bab047f92bafa6eed8d5900e9b993b