[SofaConstraint] ADD an attribute to specify the numericalTolerance.
This attribute exposes the values that were previously hardcoded in the
source code.

The default value is set to 0.0001.

Rigid3f specialization requiered some care. In previous version of the code
the Rigid3f value was set to 0.01 (without real reason).

So I revert it to 0.0001 so it may break some scene. Nevertheless a dedicated
warning message is displayed explaining to the rigid3f users that something has changed
and how they can update their scene to make it work as in the past.

The tests are also updated.

EDIT fjourdes: cherry-picked without modifications introduced in
SofaConstraint_test, which requires the messaging api in order to
compile.