UPDATE: PointSetGeometryAlgorithms. Handle the case whereby a new point has an invalid ancestor point identifier.
This case is considered as material creation, and can be used for example when you initialize a mesh starting from an empty topology. The localCoords parameter which is stored in the PointAncestorElem data structure are in this case the absolute coordinates of the point to create.

The previous implementation was only considering the case where a point is duplication of an existing point, the localCoords parameter being the offset to add to the originating point position to initialize the duplicated point coordinates.

(cherry picked from commit 9c54ed46e00f915f5bcfbdc285e2aaea9aae53e3)