UPDATE: missing EndingEvent propagation in previous unit test removing a triangle
It was not clearly appearing as a defect, a second test was thus added. In the new one, removing the ending event creates unexpected number of vertices and triangles in the information generated by Mesh2PointTopologicalMapping, showing the need to add it in the other test, as it would be the correct execution of the full topological change.