BIG WARNING: Refactoring of DataTypeInfo / AbstractTypeInfo to expose more information about data structures, including each level of containers, good support for sets and maps, planned support for structs and enums.
This is quite preliminary. Several code-breaking changes, mainly because the previous abstraction is now provided as "AbstractMultiValueTypeInfo", with other abstractions given (AbstractValueTypeInfo and AbstractContainerTypeInfo).
As a results, size / getValue / ... where renamed to be explicit which abstraction it is related:
- Single Value: resetDataValue / getDataValue / setDataValue
- Multi Value: finalSize / getFinalValue / setFinalValue
- Container: containerSize / getItemValue / editItemValue / getItemKey / editItemValue / findItem / insertItem / iterators

Also, DataTypeInfo::name() was removed in favor of existing DataTypeName::name(). As it's quite frequent that for each type only the name is redefined, it's simpler to have it in a dedicated traits class.