CLEAN: don't lock when using variable mRootTaskStatus
mRootTaskStatus should only be set from main worker thread.
The main thread should be the only one calling notifyWorkersForWork when mRootTaskStatus is nullptr. When the main thread modifies mRootTaskStatus, other threads need to be idle.
When other thread are not idle, mRootTaskStatus should be constant (except during last workUntilDone())
Because of that, we don't need to lock a mutex when checking the value of mRootTaskStatus in pushTask()

Minor, but I also made sure the mutex mWakeUpMutex was not locked during the call to notify_all, following the advice from the documentation (http://en.cppreference.com/w/cpp/thread/condition_variable/notify_all).