WARNING: Add Base::setSourceFile() and associated getters
If this component was loaded from a file, indicate from where
(this information is internal and is only meant to help developers,
do not rely on its validity as it will not be provided in all cases).

It is currently set when loading an XML file (tested) of when created from a python script (UNTESTED).