ADD: toDataRange and fromDataRange methods in DataParser
These methods operate on MultiValues and Containers and are given two
more arguments compared to toData and fromData.
The two arguments are intended to be used as iterators to only read or
write a subrange of the Data.
The new methods are optional for DataParsers and will return
unsupported_operation error if they are not implemented.