ADD: AbstractTypeInfo::getType(size_t) which reciprocates typeInfoID()
This is possible given the fact that the constructor of VirtualTypeInfo<DataType>
now fills an unordered_map with its typeInfoID and a pointer to itself.
Since there is never more than one instance of a VirtualTypeInfo<DataType> for
a given DataType, the cost of filling the map is only paid once, when
the first call to VirtualTypeInfo<DataType>::get() is made.