[SofaCore] UPDATE: ConstraintResolution postStabilize virtual method.
The default implementation of this method does not stabilize anything.
It simply clears the constraint force for the particular constraint
equation that is being solved by this ConstraintResolution instance.

Implementing the post stabilization in a concrete instance of a
ConstraintResolution usually means solving the linear part of the
constraint equation only. So for non linear ConstraintResolution like
the coulomb friction contact solving, it would mean solving the
unilateral constraint in the normal direction and set the tangential
forces to zero.

Ref T5164

(cherry picked from commit afd2065829d7d3bef9b09e6529520f97320a10dc)