[ISSofa] [SofaSimulation] FIX: guard getObject (as used by BaseContext::get<>(), getMeshTopology(), ...) against returning non-active topologies
A topology is "active" for the node it is in as well as child nodes but only if they do not contain a mapping that does not preserve topology (i.e. for which sameTopology() returns false).
Some components (such as BarycentricMapping, VisualModelImpl) for which topology events handling codes were added relatively recently were attaching to the wrong topology in many example scenes, leading to incorrect state sizes and crashes.
The "real" fix is to always use getActiveMeshTopology() when retrieving a pointer to a topology, but this first commit allows to detect all places where this can be an issue, for a relatively small performance cost.

(cherry picked from commit 48e3d74eac8f054282c8f6296c921ba4d5dc4a20)