[ISSofa] Add first implementation of the factorisation of bilateral constraints along with the dynamic matrix.
Currently the solving scheme has been only implemented by the
SparseLDLSolver.
The system matrix is augmented to take into account the jacobians
of the bilateral constraints. The system right hand side is
augmented with the value of constraint function of each bilateral
constraint.
This type of solving scheme is suitable when the bilateral constraints
express some intrinsic property of the object being modelled, and
therefore do not exhibit any redundancies, which ensures the augmented
system to still be invertible ( and positive definite ).

Implementation wise, the following was needed to be able to assemble
the augmented system:
- Bilateral constraints can specify be factored with the
  dynamics by toggling the isConstitutiveConstraint flag.
- Visitors relative to the constraint solving can inspect their
  ConstraintParams to determine which type of constraints they should
  process.

//Ref// T5705

(cherry picked from commit 657628be75802172e2ae142f05166e7e066b78c1)