[SofaHelper] Change assertion criteria to determine if a quaternion is a unit quaternion
The isUnit method can be used to determine if the quaternion we are dealing
with has still a magnitude close to one.
The threshold is now a hard value instead of a user specified epsilon.
The rationale behin this hard value can be found in the unit test
`testEpsilonSetupForUnitQuaternion`.

//Ref// T8163

(cherry picked from commit fa0edb7a50615118a2d29d519117eff71857ed5f)