[SofaCore] FIX: do not rely on "raw" static member for DataParserRegistry
static member initializations is undefined (depending on the order of the files in the command line). It is better and safer to rely on static values within static getter methods (as there the compiler guarantees that it is initialized before the first call of the method).

//Ref// T7606

(cherry picked from commit 317b013ee218bb2fb2417859c2ccfc599a3cceb3)