[ISSofa] Remove implementation of getJ in mappings (//Ref// T9010)
The rationale behind this change is that its not possible to keep the
current implementations of getJ() method and the explicit instantiation
of CompressedRowSparseMatrixMechanical types at the same time.
Some method definitions of the CompressedRowSparseMatrixMechanical
class require the matrix to be symmetric, which is usually not the
case for mappings jacobian matrices.
An alternative could have been to rely on
CompressedRowSparseMatrixConstraint to store the mapping jacobian, but
this type of matrix does not inherit from BaseMatrix and is therefore
not suitable to implement the return type of the getJ API method.

(cherry picked from commit 3749967dcf0232aa8a891364cd9616cde768b2f5)

Conflicts:
	ISSofa/modules/SofaBaseMechanics/BarycentricMapping.h
	ISSofa/modules/SofaBaseMechanics/BarycentricMapping.inl