[SofaHelper] Set FREEGLUT_LIB_PRAGMAS to 0 so that the correct lib name is used in debug mode with MSVC (//Ref// T10312)
This problem happens with "recent" versions of CMake where the debug
version of freeglut is looked for and used if it exists. For example if
its name is "freeglutd.lib" this will make the compilation fail since
the pragma to link with freeglut inside the library always look for the
"freeglut.lib" name.

(cherry picked from commit db2b809bd90ad41dfbfab34c7d7b02ae97137119)