[ISSofa] FIX: Missing method reset for ArticulatedHierarchyContainer (//Ref// T10784)
NOTE: The data 'axis' of the Articulation is initilized in the
ArticulatedHierarchyContainer and normalized in the
ArticulatedSystemMapping.
When a reset is called throughout the graph, 'axis' was not
re-initialized but was re-normalized leading to a slightly different
value due to numerical errors. This commit ensure that the reseted state
is the same than the initialized state

(cherry picked from commit b6111f7a7ff1c6a25aafc791805ab737f639ec11)